requirements.txt
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.13.3
httpx==1.0.0b0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
keyring==24.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python-okx==0.2.3
requests==2.31.0
rfc3986==1.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
zope.interface==6.1
